Change type of nota variable, for decimals acceptance

+++ src/main/java/co/udea/apifinalnote/component/nota/model/Nota.java
@@ -21,7 +21,7 @@ public class Nota {
     private Long id;
 
     @NotNull
-    private Integer nota;
+    private Double nota;
 
     @NotNull
     private Double porcentaje;
